Add JSON Web Token Auth Fixes #1

+++ backend/controllers/goalController.js
@@ -1,11 +1,12 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 
 //@desc Get all goals
 //@route GET /api/goals
 //@access Private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
   res.status(200).json(goals);
 });
 //@desc Set all goals
@@ -19,6 +20,7 @@ const setGoal = asyncHandler(async (req, res) => {
 
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
   res.status(200).json(goal);
 });
@@ -32,6 +34,18 @@ const updateGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(404);
     throw new Error("Goal not found");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  if (!user) {
+    res.status(404);
+    throw new Error("User not found");
+  }
+
+  if (goal.user.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error("Not authorized");
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(
@@ -49,12 +63,21 @@ const updateGoal = asyncHandler(async (req, res) => {
 //@access Private
 const deleteGoal = asyncHandler(async (req, res) => {
   const goal = await Goal.findById(req.params.id);
-
+  const user = await User.findById(req.user.id);
   if (!goal) {
     res.status(404);
     throw new Error("Goal not found");
   }
 
+  if (!user) {
+    res.status(404);
+    throw new Error("User not found");
+  }
+
+  if (goal.user.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error("Not authorized");
+  }
   await Goal.findByIdAndDelete(req.params.id);
 
   res.status(200).json({
